pyproject.toml
[tool.poetry]
name = "pyweb-hw-14"
version = "0.1.0"
description = ""
authors = ["Spica12 <vitaly.savenko12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.109.0"}
pytest = "^7.4.4"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg-binary = "^3.1.17"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
